added authentificatoon and edited crud
